requirements.txt
cycler==0.10.0
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
mpi4py==3.0.3
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
py-spy==0.3.6
pymp-pypi==0.4.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
sklearn==0.0
threadpoolctl==2.1.0
